Branch and Border Method
In the practical part of this work, the software for the electronic
methodic manual for the discipline "Control and Diagnostics of Information
Systems" is implemented. This software implements the solution of the problem of
building a fault finding program using the branch and boundaries method, has the
feature to enter initial data in a graphical editor and record intermediate calculations
in text form, has graphical display of the final search tree. The possibility of using
this software as an part of other software has been implemented.